Mutt fails to open PGP/Inline signed messages with \r\n EOL
Tested with Mutt 1.5.22 on Cygwin.
Problem seems to be in the file pgp.c, function pgp_application_pgp_handler. The "if (needpass == -1)" on line 511, causes that function to return -1, and Mutt then refuses to open the message, saying "Could not copy message". In debug mode, it says "Failed on attachment of type text/plain", and then "Bailing on attachment of type text/plain".